Moved all createbaseentities functions to worldobjectmanager.ts; todo: make it not slow
